ai-ml-components/requirements.txt
langchain==0.3.3
langchain-groq==0.2.0

websockets==12.*
httpx==0.*
dataclasses-json==0.*
dataclasses==0.*
typing_extensions==4.*
aenum==3.*
deprecation==2.*
aiohttp==3.*
aiofiles==23.*

python-dotenv==1.*

PyAudio==0.2.14
pydub==0.25.1 

elevenlabs==1.9.0

deepgram-sdk==3.7.3

sounddevice
